\docType{data}
\name{BGBottles}
\alias{BGBottles}
\title{Bottle data from Graybill and Burdick
#'
balanced anova data from Graybill and Burdick pg62
 Consists of 12 records of weights from 4 bottle manufacturing machines}
\format{A data frame with 35 rows and 2 variables}
\description{
  \itemize{ \item Mchine-Machine number \item weight -
  Weight of bottle produced }
}
\keyword{datasets}

