\docType{data}
\name{bull}
\alias{bull}
\title{Bull data from Graybill and Burdick
#'
unbalanced anova data from Graybill and Burdick pg69
 Consists of 35 records of Percentage yields from 6 bulls}
\format{A data frame with 35 rows and 2 variables}
\description{
  \itemize{ \item bull - Bull number \item Percentage -
  Percentage yield }
}
\keyword{datasets}

