\name{cleandf}
\alias{cleandf}
\title{Function to Strip repeat rows from a df so it looks pretty in an Xtable
Thankx to Aniko
(http://stackoverflow.com/questions/2379701/can-you-merge-cells-in-an-xtable-in-r/2380501#2380501}
\usage{
  cleandf(df, field)
}
\arguments{
  \item{df}{dataframe to be cleaned}

  \item{field}{Name of column to find duplicates with}
}
\value{
  Cleaned dataframe
}
\description{
  Function to Strip repeat rows from a df so it looks
  pretty in an Xtable Thankx to Aniko
  (http://stackoverflow.com/questions/2379701/can-you-merge-cells-in-an-xtable-in-r/2380501#2380501
}
\author{
  "Paul Hurley"
}

