\name{fsignif}
\alias{fsignif}
\title{Improved Significant Figures Function}
\usage{
  fsignif(x, digits = 3)
}
\arguments{
  \item{x}{Vector of numeric values}

  \item{digits}{Number of significant digits to display}
}
\value{
  Text string of the value to the correct sig figs
}
\description{
  Improved Significant Figures Function
}
\author{
  Paul Hurley
}

