\name{lsos}
\alias{lsos}
\title{lsos}
\usage{
  lsos(..., n = 10)
}
\arguments{
  \item{...}{}

  \item{n}{Number of Objects to List}
}
\value{
  A datafrmae listing objects
}
\description{
  List R Objects by Size
}
\details{
  Thanks to
  http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}
\author{
  gb02413
}

