\docType{data}
\name{lundanalyte}
\alias{lundanalyte}
\title{Example Dataset for Lund Outlier Test}
\format{A data frame with 18 rows and 4 variables}
\description{
  Generated analytical data Consists of 42 records for 5
  variables
}
\details{
  \itemize{ \item Analyte - Factor listing either Analyte 1
  or 2 \item Nominal - The Nominal Concentration \item
  Units - The units of concentration \item Rep - Replicate
  number of the measurement \item Actual - The actual
  reading }
}
\keyword{datasets}

