\name{pBias}
\alias{pBias}
\title{Calculate the Percentage Bias of a dataframe}
\usage{
  pBias(x, nominal, na.rm = FALSE)
}
\arguments{
  \item{x}{a value}

  \item{nominal}{the nominal value}
}
\value{
  numeric Return the percentage bias of x in terms of
  nominal
}
\description{
  Calculate the Percentage Bias of a dataframe
}
\author{
  Paul Hurley
}

