\docType{data}
\name{phosphorus}
\alias{phosphorus}
\title{Example Dataset for Lund Outlier Test}
\format{A data frame with 18 rows and 4 variables}
\description{
  Dataset from the original Lund papr Consists of 18
  records for two variables and one response
}
\details{
  \itemize{ \item Y - Response \item X1 - Variable 1 \item
  X2 - Variable 2 }
}
\keyword{datasets}

