\name{cleanf}
\alias{cleanf}
\title{Function to Strip repeat values from a df so it looks pretty in an Xtable
Thankx to Aniko
(http://stackoverflow.com/questions/2379701/can-you-merge-cells-in-an-xtable-in-r/2380501#2380501}
\usage{
  cleanf(x)
}
\arguments{
  \item{dataframe}{to be cleaned}
}
\value{
  Cleaned dataframe
}
\description{
  Function to Strip repeat values from a df so it looks
  pretty in an Xtable Thankx to Aniko
  (http://stackoverflow.com/questions/2379701/can-you-merge-cells-in-an-xtable-in-r/2380501#2380501
}
\author{
  "Paul Hurley"
}

