\name{pastewrap}
\alias{pastewrap}
\title{Insert linebreaks into long strings}
\usage{
  sapply(foo, pastwrap, len=22)
}
\arguments{
  \item{x}{A string}

  \item{len}{Length to break at}
}
\value{
  string with linebreaks inserted every len characters
}
\description{
  Insert linebreaks into long strings
}
\author{
  Paul Hurley
}

